feat(mqtt): add apigear client and server adapters
Apigear goal is to use Mqtt as an rpc for the objext. It splits to a
client side of object and service side object.
this commit adapts the mqtt client (the BaseClient) to those apigeaar objects:
a remote object client side (Client) and apigear remote server
side (Service)
They add the methods that describe required functionality, but not all
of them are yet implemented.
Those are not yet the inteface taiored classes yet - those are the network
and general mqtt handling classes