feat(olink): improve testing
Use test_value filter to provide non deafault values for olink tests.
With a request of setting a property diffrent than actual, the olink
client actually sends a request and allows to be checked if is working.
Also non default values were used as signal parameters, and checked if
properly recieived on client side.
Ticket #41